refactor: inline trivial gateway/{channels,api,spaces} index re-exports
Each of these index.ts files re-exported a single symbol from one neighbour
module, adding an unnecessary indirection layer. Importers now reach for the
implementation file directly.
