--- packages/server/src/gateway/api/index.ts
@@ -1,1 +1,0 @@
-export { ApiPlatform, type ApiPlatformConfig } from "./platform.js";
