--- packages/server/src/gateway/channels/index.ts
@@ -1,1 +1,0 @@
-export { ChannelBindingService } from "./binding-service.js";
