+++ packages/server/src/gateway/platform.ts
@@ -12,7 +12,7 @@ import type { ProviderOAuthStateStore } from "./auth/oauth/state-store.js";
 import type { AgentSettingsStore } from "./auth/settings/agent-settings-store.js";
 import type { ModelPreferenceStore } from "./auth/settings/model-preference-store.js";
 import type { UserAgentsStore } from "./auth/user-agents-store.js";
-import type { ChannelBindingService } from "./channels/index.js";
+import type { ChannelBindingService } from "./channels/binding-service.js";
 import type { ArtifactStore } from "./files/artifact-store.js";
 import type { WorkerGateway } from "./gateway/index.js";
 import type {
