+++ packages/server/src/gateway/routes/public/agents.ts
@@ -18,7 +18,7 @@ import type {
 import { buildDefaultSettingsFromSource } from "../../auth/settings/template-utils.js";
 import type { SettingsTokenPayload } from "../../auth/settings/token-service.js";
 import type { UserAgentsStore } from "../../auth/user-agents-store.js";
-import type { ChannelBindingService } from "../../channels/index.js";
+import type { ChannelBindingService } from "../../channels/binding-service.js";
 import {
   resolveSettingsLookupUserId,
   verifyOwnedAgentAccess,
