--- packages/server/src/gateway/spaces/index.ts
@@ -1,1 +1,0 @@
-export * from "./space-resolver.js";
